requirements.txt
py-trees
google-generativeai
vosk
pyaudio
python-dotenv
google-cloud-texttospeech
google-cloud-speech
